+++ internal/providers/redis.go
@@ -61,7 +61,7 @@ func (c *Redis) BatchGet(keys []string) (map[string]string, error) {
 }
 
 // BatchPut сохраняет несколько значений за один запрос
-func (c *Redis) BatchPut(items map[string]string, ttls map[string]uint) error {
+func (c *Redis) BatchPut(items map[string]string, ttls map[string]int64) error {
 	if len(items) == 0 {
 		return nil
 	}
